Rename `src/` and add `__init__.py`
